movies_admin/requirements.txt
django-admin == 2.0.2
django == 4.2.5
python-dotenv == 1.0.1
django-split-settings == 1.2.0
django-debug-toolbar == 4.3.0
psycopg2-binary
psycopg == 3.1.17
django-health-check == 3.18.1
gunicorn==20.1.0
django-cors-headers==4.3.1
tenacity==8.2.3
elasticsearch==8.12.1
pydantic==2.6.3
pydantic-settings==2.2.1
redis==5.0.2